Implement ErrorFlash modal and logic
Attempted to use session.flah but I could not get it working correctly
with JS and no-js component.

So, we use a normal session key with some
remix endpoints to clear it and read it.

This means that it works and dismisses in both js
and no-js scenarios but it cannot
be auto dismissed when changing routes.